fixed users.js
